fix: bone weight normalize

+++ src/LibGLaDOS/core/loader/AssimpLoader.h
@@ -20,7 +20,7 @@ typedef aiMatrix4x4t<ai_real> aiMatrix4x4;
 
 namespace GLaDOS {
     struct BoneInfo {
-        int8_t id;
+        int32_t id;
         std::string name;
         Mat4<real> offset;
     };
@@ -50,7 +50,7 @@ namespace GLaDOS {
         bool loadNode(aiNode* node, const aiScene* scene);
         Mesh* loadMesh(aiMesh* mesh);
         void parseBoneWeight(Vector<Vertex>& vertices, aiBone* bone);
-        int8_t findOrCacheBone(const std::string& name, aiBone* bone);
+        int32_t findOrCacheBone(const std::string& name, aiBone* bone);
         Texture* loadTexture(aiMaterial* material, aiTextureType textureType);
         AnimationClip* loadAnimation(aiAnimation* animation);
